+++ HISTORY.rst
@@ -6,7 +6,7 @@ History
 1.0.0 (2020-06-18)
 +++++++++++++++++++
 
-* Adds support for keying batch geocode results (thanks liufran1 and Unix-Code!)
+* Adds support for keying batch geocode and reverse geocode results (thanks liufran1 and Unix-Code!)
 
 0.12.0 (2020-06-04)
 +++++++++++++++++++